requirements.txt
streamlit
numpy
plotly
reportlab
Pillow
